restructure iomod guest code
