--- backends/aws-lambda/iomod/guest/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "assemblylift-awslambda-iomod-guest"
-version = "0.1.1"
-description = "AssemblyLift AWS Lambda IOmod guest library"
-authors = ["Akkoro and the AssemblyLift contributors <assemblylift@akkoro.io>"]
-edition = "2018"
-license-file = "../../../../LICENSE.md"
-repository = "https://github.com/akkoro/assemblylift"
-readme = "README.md"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = "1.0.106"
-serde_json = "1.0.53"
-bytes = "0.5.4"
-base64 = "0.12.1"
-paste = "0.1.12"
-assemblylift_core_guest = { version = "0.1.0", package = "assemblylift-core-guest", path = "../../../../core/guest" }
-assemblylift_core_event_guest = { version = "0.1.1", package = "assemblylift-core-event-guest", path = "../../../../core/event/guest" }
