--- backends/aws-lambda/iomod/guest/README.md
@@ -1,4 +1,0 @@
-assemblylift-awslambda-iomod-guest
----------------------------
-
-AssemblyLift AWS Lambda IOmod guest library.
