+++ backends/aws-lambda/iomod/plugins/dynamodb/guest/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "guest"
+name = "assemblylift-awslambda-iomod-plugin-dynamodb-guest"
 version = "0.1.0"
 authors = ["xlem <xlem@akkoro.io>"]
 edition = "2018"
@@ -7,3 +7,12 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+serde = "1.0.106"
+serde_json = "1.0.53"
+bytes = "0.5.4"
+base64 = "0.12.1"
+paste = "0.1.12"
+
+assemblylift_core_guest = { version = "0.1.0", package = "assemblylift-core-guest", path = "../../../../../../core/guest" }
+assemblylift_core_event_guest = { version = "0.1.1", package = "assemblylift-core-event-guest", path = "../../../../../../core/event/guest" }
+assemblylift_core_iomod_guest = { version = "0.1.0", package = "assemblylift-core-iomod-guest", path = "../../../../../../core/iomod/guest" }
