+++ backends/aws-lambda/iomod/plugins/dynamodb/guest/src/serialization.rs
