+++ backends/aws-lambda/iomod/plugins/dynamodb/guest/src/structs.rs
