+++ core/iomod/guest/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "assemblylift-core-iomod-guest"
+version = "0.1.0"
+authors = ["xlem <xlem@akkoro.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
